dev-2752: move cgacs/frecs back to references
