--- usaspending_api/agencies/models/__init__.py
@@ -1,5 +1,0 @@
-from usaspending_api.agencies.models.cgac import CGAC
-from usaspending_api.agencies.models.frec import FREC
-
-
-__all__ = ["CGAC", "FREC"]
