+++ usaspending_api/common/etl/primatives.py
@@ -200,7 +200,7 @@ def wrap_dblink_query(
 ) -> Composed:
     """ Wraps a query in a dblink compatible query so that it can be run on a remote server. """
     inner_sql = convert_composable_query_to_string(sql)
-    select_columns = make_column_list(columns, "r")
+    select_columns = make_column_list(columns, alias)
     typed_columns = make_typed_column_list(columns, data_types)
     sql = """
         select {select_columns}
