+++ usaspending_api/references/models/__init__.py
@@ -1,7 +1,9 @@
 from usaspending_api.references.models.agency import Agency
 from usaspending_api.references.models.cfda import Cfda
+from usaspending_api.references.models.cgac import CGAC
 from usaspending_api.references.models.definition import Definition
 from usaspending_api.references.models.filter_hash import FilterHash
+from usaspending_api.references.models.frec import FREC
 from usaspending_api.references.models.frec_map import FrecMap
 from usaspending_api.references.models.gtas_total_obligation import GTASTotalObligation
 from usaspending_api.references.models.legal_entity import LegalEntity
@@ -20,8 +22,10 @@ from usaspending_api.references.models.toptier_agency import ToptierAgency
 __all__ = [
     "Agency",
     "Cfda",
+    "CGAC",
     "Definition",
     "FilterHash",
+    "FREC",
     "FrecMap",
     "GTASTotalObligation",
     "LegalEntity",
