+++ usaspending_api/references/tests/test_tas_autocomplete.py
@@ -9,8 +9,8 @@ BASE_ENDPOINT = "/api/v2/autocomplete/accounts/"
 
 @pytest.fixture
 def test_data(db):
-    mommy.make("agencies.CGAC", cgac_code="000", agency_name="Agency 000", agency_abbreviation="A000")
-    mommy.make("agencies.CGAC", cgac_code="002", agency_name="Agency 002", agency_abbreviation="A002")
+    mommy.make("references.CGAC", cgac_code="000", agency_name="Agency 000", agency_abbreviation="A000")
+    mommy.make("references.CGAC", cgac_code="002", agency_name="Agency 002", agency_abbreviation="A002")
 
     award = mommy.make("awards.Award")
